Added screen to search for city by name
City search is performed via the OpenWeatherMap API and displayed to the user
The API key is stored in the Config.plist list. This file must be created before the project is compiled. It is omitted to ensure the API key is not made public
Modified the Info.plist file to allow insecure HTTP connections to api.openweathermap.org due to the API not having HTTPS support